Fixes from PR-5
