+++ flightmate/src/main/java/com/flightmate/libs/Route.java
@@ -7,4 +7,5 @@ public class Route {
 	public static String SIGNUP = "signup.jsp";
 	public static String AIRPORT = "airport.jsp";
 	public static String AIRCRAFT = "aircraft.jsp";
+	public static String FLIGHT = "flight.jsp";
 }
