+++ flightmate/src/main/java/com/flightmate/servlets/AircraftServlet.java
@@ -3,6 +3,7 @@ package com.flightmate.servlets;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,11 +12,14 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.flightmate.beans.Aircraft;
+import com.flightmate.beans.Airport;
 import com.flightmate.beans.User;
 import com.flightmate.dao.AircraftDao;
 import com.flightmate.dao.AirportDao; // For fetching airports
+import com.flightmate.dao.UserDao;
 import com.flightmate.libs.Role;
 import com.flightmate.libs.Route;
+import com.flightmate.libs.builders.AircraftBuilder;
 import com.flightmate.libs.services.SessionService;
 
 @WebServlet("/aircraft")
@@ -30,8 +34,12 @@ public class AircraftServlet extends HttpServlet {
             resp.sendRedirect(Route.LOGIN);
             return;
         } 
+        
+        List<User> administrators = UserDao.getDao().getAllAdministrators();
+        List<Airport> airports = AirportDao.getAllAirports();
 
-        req.setAttribute("airports", AirportDao.getDao().getAllAirports());
+        req.setAttribute("administrators", administrators);
+        req.setAttribute("airports", airports);
         req.getRequestDispatcher(Route.AIRCRAFT).forward(req, resp);
     }
 
@@ -44,17 +52,33 @@ public class AircraftServlet extends HttpServlet {
             return;
         }
 
-        String model = req.getParameter("model");
-        String details = req.getParameter("details");
-        int airportId = Integer.parseInt(req.getParameter("airport"));
+        String aircraftModel = req.getParameter("aircraftModel");
+        LocalDate manufactureDate = LocalDate.parse(req.getParameter("manufactureDate"));
+        LocalDate lastMaintenanceDate = LocalDate.parse(req.getParameter("lastMaintenanceDate"));
+        LocalDate nextMaintenanceDate = LocalDate.parse(req.getParameter("nextMaintenanceDate"));
+        String aircraftNotes = req.getParameter("aircraftNotes");
+        int administratorId = Integer.parseInt(req.getParameter("administratorId"));
+        int airportId = Integer.parseInt(req.getParameter("airportId"));
 
-        Aircraft aircraft = new Aircraft(0, model, LocalDate.now(), LocalDate.now(), LocalDate.now(), details, LocalDateTime.now(), airportId, airportId);
-        boolean success = AircraftDao.getDao().addAircraft(aircraft);
+
+        Aircraft aircraft = new AircraftBuilder()
+                .setAircraftModel(aircraftModel)
+                .setManufactureDate(manufactureDate)
+                .setLastMaintenanceDate(lastMaintenanceDate)
+                .setNextMaintenanceDate(nextMaintenanceDate)
+                .setAircraftNotes(aircraftNotes)
+                .setCreatedAt(LocalDateTime.now())
+                .setAdministratorId(administratorId)
+                .setAirportId(airportId)
+                .create();
+
+        // Save the Aircraft to the database
+        boolean success = AircraftDao.addAircraft(aircraft);
 
         if (success) {
             req.setAttribute("success", "Aircraft added successfully!");
         } else {
-            req.setAttribute("message", "Failed to add aircraft. Please try again.");
+            req.setAttribute("error", "Failed to add aircraft.");
         }
 
         doGet(req, resp);
