+++ flightmate/src/main/java/com/flightmate/servlets/DashboardServlet.java
@@ -9,6 +9,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.flightmate.beans.User;
+import com.flightmate.dao.UserDao;
+import com.flightmate.libs.Role;
 import com.flightmate.libs.Route;
 import com.flightmate.libs.services.SessionService;
 
@@ -24,6 +26,28 @@ public class DashboardServlet extends HttpServlet {
 			resp.sendRedirect(Route.LOGIN);
 			return;
 		}
+		
+		String action = req.getParameter("action");
+        if (action != null && user.getRole().equals(Role.ADMINISTRATOR)) {
+            try {
+                switch (action) {
+                    case "edit":
+                        handleEdit(req, resp); 
+                        return;
+                    case "delete":
+                        handleDelete(req);
+                        break;
+                    default:
+                        req.setAttribute("error", "Invalid action.");
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+                req.setAttribute("error", "Error processing action: " + e.getMessage());
+            }
+        }
+        
+		req.setAttribute("users", UserDao.getDao().getAllUsers()); 
+//      req.setAttribute("flights", FlightDao.getDao().getAllFlights()); 
 						
 		req.getRequestDispatcher(Route.DASHBOARD).forward(req, resp);
 	}
@@ -37,6 +61,82 @@ public class DashboardServlet extends HttpServlet {
 			return;
 		}
 		
+		String action = req.getParameter("action");
+		if ("save".equals(action)) {
+            handleSave(req, resp);
+            
+        }
+		
 		doGet(req, resp);
 	}
+	
+	
+	private void handleEdit(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        String userIdParam = req.getParameter("id");
+        if (userIdParam == null || userIdParam.isEmpty()) {
+            req.setAttribute("error", "User ID is required for editing.");
+            return;
+        }
+
+        try {
+            int userId = Integer.parseInt(userIdParam);
+
+            User user = UserDao.getDao().getUserById(userId);
+            if (user == null) {
+                req.setAttribute("error", "User not found.");
+                return;
+            }
+
+            req.setAttribute("user", user);
+            req.getRequestDispatcher("edit_user.jsp").forward(req, resp);
+        } catch (NumberFormatException e) {
+            req.setAttribute("error", "Invalid User ID format.");
+        }
+    }
+
+    private void handleDelete(HttpServletRequest req) {
+        String userIdParam = req.getParameter("id");
+        if (userIdParam == null || userIdParam.isEmpty()) {
+            req.setAttribute("error", "User ID is required for deletion.");
+            return;
+        }
+
+        try {
+            int userId = Integer.parseInt(userIdParam);
+
+            boolean deleted = UserDao.getDao().deleteUser(userId);
+
+            if (deleted) {
+                req.setAttribute("success", "User deleted successfully.");
+            } else {
+                req.setAttribute("error", "Failed to delete user.");
+            }
+        } catch (NumberFormatException e) {
+            req.setAttribute("error", "Invalid User ID format.");
+        }
+    }
+    
+    private void handleSave(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        String userIdParam = req.getParameter("user_id");
+        String firstName = req.getParameter("first_name");
+        String lastName = req.getParameter("last_name");
+        String email = req.getParameter("email_address");
+        String roleIdParam = req.getParameter("role_id");
+
+        try {
+            int userId = Integer.parseInt(userIdParam);
+            int roleId = Integer.parseInt(roleIdParam);
+
+            boolean updated = UserDao.getDao().updateUser(firstName, lastName, email, roleId);
+
+            if (updated) {
+                req.setAttribute("success", "User updated successfully.");
+            } else {
+                req.setAttribute("error", "Failed to update user.");
+            }
+        } catch (NumberFormatException e) {
+            req.setAttribute("error", "Invalid input for User ID or Role ID.");
+        }
+
+    }
 }
